refactor(daemon): split compile_journal.rs into focused submodules
Convert crates/zccache-daemon/src/compile_journal.rs (2,072 LOC) into a
compile_journal/ module directory with 4 source files and 5 test files.
Each new file is well under 1,000 LOC (largest: tests/entry.rs at 454).
No behavior change — purely a code split. Public API surface is identical:
mod.rs re-exports derive_*/extract_outcome so every existing
`zccache_daemon::compile_journal::<item>` path stays reachable.

Production code layout:
- mod.rs: pub types (JournalEntry, MissDiff, SelfProfileNs,
  SelfProfileSpans, JournalContext, CompileJournal) + pub mod miss_reason
- derive.rs: derive_crate_name/type/output_ext (pure argv parsers)
- outcome.rs: extract_outcome (Response -> journal-tuple)
- journal_thread.rs: JournalMessage, journal_thread, rotate_journal,
  gc_journal_files, JOURNAL_MAX_SIZE/JOURNAL_MAX_FILES

Test layout (tests/):
- mod.rs: wait_for_lines + legacy_entry + make_ctx helpers
- derive.rs: 11 tests for the argv parsers
- outcome.rs: 16 tests for extract_outcome (basic + edge + miss_reason)
- miss_reason.rs: 9 tests for the closed-enum constants, MissDiff
  serialization, and end-to-end JSONL miss-reason wiring
- entry.rs: 17 tests for JournalEntry serialization (legacy + profile
  fields) + SelfProfileSpans / with_profile_fields
- journal_file.rs: 14 tests for CompileJournal end-to-end file writes,
  per-session journals, close/reopen, concurrent logging, rotation, GC

All 226 daemon lib unit tests still pass (10 ignored — unchanged from
the baseline on main). Clippy clean on workspace. Doc clean.
